Extract weekend truce time utilities
